+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion '26.0.2'
     defaultConfig {
-        applicationId "nriviere97.auxduty"
+        applicationId "ca.riverboatmedia.auxduty"
         minSdkVersion 15
         targetSdkVersion 25
         versionCode 1
