+++ app/src/main/java/ca/riverboatmedia/auxduty/positionCallback.java
@@ -1,0 +1,31 @@
+package ca.riverboatmedia.auxduty;
+
+import android.content.Context;
+import android.util.Log;
+import android.view.View;
+import android.widget.ListView;
+
+import java.util.HashMap;
+
+/**
+ * Created by briviere on 9/24/17.
+ */
+
+public class positionCallback {
+    private Context mContext;
+    ListView mList;
+    HashMap mMap;
+    positionCallback(Context context, ListView lv, HashMap map) {
+        mContext = context;
+        mList = lv;
+        mMap = map;
+    }
+
+    public void call(Integer pos, View view, boolean inflated) {
+        if (inflated) {
+            mMap.put(pos, pos);
+        } else {
+            mMap.put(pos, mList.getPositionForView(view));
+        }
+    }
+}
