--- app/src/main/java/com/example/auxduty/ListViewLoader.java
@@ -1,71 +1,0 @@
-package nriviere97.auxduty;
-
-import android.app.ListActivity;
-import android.app.LoaderManager;
-import android.content.CursorLoader;
-import android.content.Loader;
-import android.database.Cursor;
-import android.os.Bundle;
-import android.provider.ContactsContract;
-import android.provider.MediaStore;
-import android.view.View;
-import android.widget.SimpleCursorAdapter;
-import android.widget.TextView;
-
-import nriviere97.auxduty.data.musicDataContract.MusicEntry;
-
-public class ListViewLoader extends ListActivity implements LoaderManager.LoaderCallbacks<Cursor> {
-
-    // This is the Adapter being used to display the list's data
-    SimpleCursorAdapter mAdapter;
-
-    // These are the Contacts rows that we will retrieve
-    static final String[] PROJECTION = new String[] {MusicEntry.COLUMN_SONG_NAME, MusicEntry.COLUMN_ARTIST};
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        // For the cursor adapter, specify which columns go into which views
-        String[] fromColumns = {MusicEntry.COLUMN_SONG_NAME, MusicEntry.COLUMN_ARTIST };
-
-        int[] toViews = {(R.id.song_id), (R.id.artist_id)};
-        // The TextView in simple_list_item_1
-        // Create an empty adapter we will use to display the loaded data.
-        // We pass null for the cursor, then update it in onLoadFinished()
-        mAdapter = new SimpleCursorAdapter(this, android.R.layout.simple_list_item_1, null,
-                fromColumns, toViews, 0);
-        setListAdapter(mAdapter);
-
-        // Prepare the loader.  Either re-connect with an existing one,
-        // or start a new one.
-        getLoaderManager().initLoader(0, null, this);
-    }
-
-    // Called when a new Loader needs to be created
-    public Loader<Cursor> onCreateLoader(int id, Bundle args) {
-        // Now create and return a CursorLoader that will take care of
-        // creating a Cursor for the data being displayed.
-        return new CursorLoader(this, MediaStore.Audio.Media.EXTERNAL_CONTENT_URI,
-                PROJECTION, null, null, null);
-    }
-
-    @Override
-    // Called when a previously created loader has finished loading
-    public void onLoadFinished(Loader<Cursor> loader, Cursor data) {
-        // Swap the new cursor in.  (The framework will take care of closing the
-        // old cursor once we return.)
-        mAdapter.swapCursor(data);
-    }
-
-    @Override
-    // Called when a previously created loader is reset, making the data unavailable
-    public void onLoaderReset(Loader<Cursor> loader) {
-        // This is called when the last Cursor provided to onLoadFinished()
-        // above is about to be closed.  We need to make sure we are no
-        // longer using it.
-        mAdapter.swapCursor(null);
-    }
-}
-
-
