+++ app/src/main/res/layout/activity_join_session.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="nriviere97.auxduty.joinSession"
+    tools:context="ca.riverboatmedia.auxduty.joinSession"
     tools:ignore="ExtraText">
 
     <LinearLayout
