+++ app/src/main/res/layout/activity_main_screen.xml
@@ -6,7 +6,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="nriviere97.auxduty.MainScreen"
+    tools:context="ca.riverboatmedia.auxduty.MainScreen"
     tools:ignore="ExtraText">
 
     <ImageView android:layout_height="wrap_content"
